lib: refacto SCLang experimental bindings -> first rename
