+++ renardo_lib/renardo_lib/Effects/NewEffects.py
@@ -1,6 +1,6 @@
 import inspect
 
-from renardo_lib.SCLang.SCLang import cls, instance, Vibrato, RHPF
+from renardo_lib.SCLang.SCLangExperimentalBindings import cls, instance, Vibrato, RHPF
 
 # To be moved to SCLang later
