+++ renardo_lib/renardo_lib/SCLang/Env.py
@@ -1,4 +1,4 @@
-from renardo_lib.SCLang.SCLang import instance, cls
+from renardo_lib.SCLang.SCLangExperimentalBindings import instance, cls
 
 class EnvGen(instance):
     value = "Env"
