+++ renardo_lib/renardo_lib/SCLang/__init__.py
@@ -1,4 +1,4 @@
-from renardo_lib.SCLang.SCLang import *
+from renardo_lib.SCLang.SCLangExperimentalBindings import *
 from renardo_lib.SCLang.SynthDef import SynthDefs, SynthDef, SynthDefProxy, SampleSynthDef, CompiledSynthDef
 # from renardo_lib.SCLang import Env
